requirements.txt
fastapi
uvicorn
pymupdf
python-docx
python-dotenv
python-multipart
requests